==> code/8-Report-impact3.Rmd <==
---
output: 
  pdf_document: 
    fig_caption: yes
    number_sections: yes
    toc: yes
    toc_depth: 3
---
```{r , echo=FALSE, warning=FALSE, message=FALSE, cache=TRUE}
## Sys.setenv(PATH = paste(Sys.getenv("PATH"), "C:\\R\\miketex\\miktex\\bin\\x64 ", sep=.Platform$path.sep))

ctrname <- "Tunisia"
yearreport <- "2017"


#Sys.which("pdflatex")
#Sys.getenv("PATH")
```

---
title: "Analysis of `r ctrname` Country Operation Plan  - Impact indicators for `r yearreport`"
author: "MENA Protection Services - Information Management"
date: "`r Sys.Date()`"
---


```{r include=FALSE, cache=FALSE}

library(ggplot2)
library(ggthemes)
library(rmarkdown)
library(xtable)
library(knitr)
library(pander)  ## library to format regression output in Rmd

```


```{r , echo=FALSE, warning=FALSE, message=FALSE, cache=TRUE}
setwd("D:/R-project/focus-analysis") 
focus.impact <- read.csv("data/impact.csv")
#names(focus.impact)
## load all packages
source("D:/R-project/focus-analysis/code/0-package.R")

```


```{r , echo=FALSE, warning=FALSE, message=FALSE, cache=FALSE}

library(reshape2)
## Subset for the country
focus1.impact <-focus.impact[focus.impact$operationName == ctrname, ]
source("D:/R-project/focus-analysis/code/timeindic.R")

```

*** 

# Introduction 

UNHCR [FOCUS](http://focus.unhcr.org) system allows for a consistent organisation of all information related to UNHCR programme. The data used for the analysis has been pulled out of Focus Reader. Indicator were also analysed in relation with the Budget reporting managed in MSRP and exposed through Global Focus Insight.

This report allow to provide more analytical insight from the profile of the operation plan for each country. It is automatically generated from the data extracted using a standard analysis script in the R statistical language. If you identify errors or additional interesting analysis to be added, please contact [Edouard Legoupil, snr IMO, MENA Protection Services](mailto:legoupil@unhcr.org?Subject=FocusAnalysisReport).


## Identification of information source and line of responsibility for the indicator monitoring plan

Since 2016, Preparation of monitoring plan of impact and performance indicators has become compulsory (cf. [UNHCR Programme Manual”. page 187](https://intranet.unhcr.org/content/dam/UNHCR/dpsm/pass/pm/PM%20-%20PART%20II%20-%20MONITORING.pdf)).

### Establish line of responsibility
The UNHCR RBM (Result Based Management) logical framework has been here remapped to the standard job profile present within operations  (i.e. Registration, SGBV, CP, RST, RSD, etc.) .


### Sourcing indicators
The following provided suggested sources for each indicator per technical area of responsibility.
To do so, all the indicators from the UNHCR Result based Framework were mapped to one of the following data source: 

 1. National Statistical office (or Stats from Ministry – Education – Labour – Interior…) 
 2. Registration 
 3. Case management & assistance tracking – Interagency 
 4. Record Tracking Sheet – linked to one agency – potential double count – data protection risk… 
 5. Representative Household survey 
 6. Protection Monitoring 
 7. Expert Panel (for instance JAM) 
 8. Structured community discussions 
 9. Key Informant 
 10. Registry of official document (SOPs, Agreement, Laws..) 
 
When impact indicators can be selected from more than one available source wihtin the country, the following considerations should be take in account the decide on the one to retain:  

 * __Authority__ (for instance, National Stats always to come first)
 * __Accuracy__ (household survey better than Key informant)
 * __Convenience__ (for instance expert panel easier to organize than key Informant)
 * __Economies of scale__ (for instance one unique HH survey can help to cover multiple indicators)
 * __Cost__ (representative household survey ~ minimum 25 K$…)


## Support the indicator review excercise

An important element while developing the monitoring plan is to consider what is the importance of the indicator versus the burden 9in terms of workload) and the cost of obtaining the data? For instance, if the indicator has been part of the framework last year but only a few partners used or only a few values were collected throughout the year, is it worth keeping it?


### Missing values 
 
The most common data quality issues for indicators are the following: 

 * "No Baseline"
 * "No Target"
 * "No Mid Year"

Those indicators becomes useless in terms of planning as there won't be anyway to use them for achievement level measurement. The analysis of indicator levels can be done in different ways dpending on whether the difference is looking at: 
 * End year Value or mid year Report;
 * Target Value as defined per operating level (all needs) and target Value as defined per Operating Plan (prioritised needs);
 * Achievement in terms of progress from the base, progres toward the target or absolute progres between the base and the target value. 

### Analysis Classes 

The analysis of indicators difference is then classified along the following classes:

 * "No Mid Year review"
 * "No End Year report"
 * "No Baseline"
 * "No OP Target"
 * "No OL Target"
 * "Underplan" - difference is negative
 * "Underperf" - difference below 30% (for mid year) or 50% (for End  year)
 * "Normal" - difference beetween 30% (for mid year) or 50% (for End  year) & 70 % (for mid year) or 90% (for End  year)
 * "Overperf"- difference is over 70 % (for mid year) or 90% (for End  year)
 * "Overplan" - difference is over 100 %


### Assess the consistency of indicators selection

The second important element is to consider  the consistency within the selection of indicators in terms of longitudinal analysis.

###  Indicator consistency over time and accross the subregion

An element of decision for indicator selection is to assess how the selection of indicators match the most used indicators in the region over the recent years. 

### Number of indicators per output 

When defining impact indicators, (the “excess luggage”)has been matched to more than one indicator  

### Comparison of baseline value over years

The objective is to review the consistency of baseline value between previous year. 
if current baseline is missing and the same indicator was used in the previous year



##  Indicator in relation with Strategies & Global priorities
The second important element is to consider  the consistency within the selection of indicators  in terms of compliance with global recommandations.

### Global Strategic Priorities (GSP)

### SGBV Updated Global Strategy 

### Global Child Protection Framework 

### Global Education Strategy

### Community mobilization Strategy

### Global Strategy for Livelihoods

### Global Strategy for Safe Access to Fuel and Energy (SAFE)



*** 
 
#  Refugee Registration
## Source Information
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis'}
this.subtype <- "Registration"
source("D:/R-project/focus-analysis/code/revindic.R")
```

## Review indicators
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis', fig.width = 10, fig.height = 7}
this.subtype <- "Registration"
source("D:/R-project/focus-analysis/code/plotindic.R")
```

## Check with global recommandations

***  
#  Refugee Status Determination (RSD)
## Source Information
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis'}
this.subtype <- "RSD"
source("D:/R-project/focus-analysis/code/revindic.R")
```

## Review indicators
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis', fig.width = 10, fig.height = 7}
this.subtype <- "RSD"
source("D:/R-project/focus-analysis/code/plotindic.R")
```

## Check with global recommandations


***  
#  Resettlement
## Source Information
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis'}
this.subtype <- "Resettlement"
source("D:/R-project/focus-analysis/code/revindic.R")
```

## Review indicators
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis', fig.width = 10, fig.height = 7}
this.subtype <- "Resettlement"
source("D:/R-project/focus-analysis/code/plotindic.R")
```

## Check with global recommandations

*** 
#  Child Protection
## Source Information
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis'}
this.subtype <- "ChildProtection"
source("D:/R-project/focus-analysis/code/revindic.R")
```

## Review indicators
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis', fig.width = 10, fig.height = 7}
this.subtype <- "ChildProtection"
source("D:/R-project/focus-analysis/code/plotindic.R")
```

## Check with global recommandations

*** 
#  Sexual and other forms of gender-based violence (SGBV)
## Source Information
 
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis'}
this.subtype <- "SGBV"
source("D:/R-project/focus-analysis/code/revindic.R")
```

## Review indicators
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis', fig.width = 10, fig.height = 7}
this.subtype <- "SGBV"
source("D:/R-project/focus-analysis/code/plotindic.R")
```

## Check with global recommandations

***
#  Education
## Source Information
 
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis'}
this.subtype <- "Education"
source("D:/R-project/focus-analysis/code/revindic.R")
```

## Review indicators
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis', fig.width = 10, fig.height = 7}
this.subtype <- "Education"
source("D:/R-project/focus-analysis/code/plotindic.R")
```

## Check with global recommandations

***
#  Legal issues
## Source Information
 
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis'}
this.subtype <- "Legal"
source("D:/R-project/focus-analysis/code/revindic.R")
```

## Review indicators
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis', fig.width = 10, fig.height = 7}
this.subtype <- "Legal"
source("D:/R-project/focus-analysis/code/plotindic.R")
```

## Check with global recommandations

*** 
# Mixed Migration
## Source Information
 
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis'}
this.subtype <- "Mixed migration"
source("D:/R-project/focus-analysis/code/revindic.R")
```

## Review indicators
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis', fig.width = 10, fig.height = 7}
this.subtype <- "Mixed migration"
source("D:/R-project/focus-analysis/code/plotindic.R")
```

## Check with global recommandations

*** 
#  Statelessness
## Source Information
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis'}
this.subtype <- "Statelessness"
source("D:/R-project/focus-analysis/code/revindic.R")
```

## Review indicators
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis', fig.width = 10, fig.height = 7}
this.subtype <- "Statelessness"
source("D:/R-project/focus-analysis/code/plotindic.R")
```

## Check with global recommandations

*** 
#  Community Services
## Source Information
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis'}
this.subtype <- "Community Services"
source("D:/R-project/focus-analysis/code/revindic.R")
```

## Review indicators
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis', fig.width = 10, fig.height = 7}
this.subtype <- "Community Services"
source("D:/R-project/focus-analysis/code/plotindic.R")
```

## Check with global recommandations

***
#  Community Based Protection
## Source Information
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis'}
this.subtype <- "Community Based Protection"
source("D:/R-project/focus-analysis/code/revindic.R")
```

## Review indicators
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis', fig.width = 10, fig.height = 7}
this.subtype <- "Community Based Protection"
source("D:/R-project/focus-analysis/code/plotindic.R")
```

## Check with global recommandations

***
#  Communication
## Source Information
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis'}
this.subtype <- "Communication"
source("D:/R-project/focus-analysis/code/revindic.R")
```

## Review indicators
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis', fig.width = 10, fig.height = 7}
this.subtype <- "Communication"
source("D:/R-project/focus-analysis/code/plotindic.R")
```

## Check with global recommandations

*** 
 
#  Protection Coordination
## Source Information
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis'}
this.subtype <- "ProtCoordination"
source("D:/R-project/focus-analysis/code/revindic.R")
```

## Review indicators
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis', fig.width = 10, fig.height = 7}
this.subtype <- "ProtCoordination"
source("D:/R-project/focus-analysis/code/plotindic.R")
```

## Check with global recommandations

*** 
#  Information Management
## Source Information
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis'}
this.subtype <- "Information Management"
source("D:/R-project/focus-analysis/code/revindic.R")
```

## Review indicators
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis', fig.width = 10, fig.height = 7}
this.subtype <- "Information Management"
source("D:/R-project/focus-analysis/code/plotindic.R")
```

## Check with global recommandations

***

# Health
## Source Information
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis'}
this.subtype <- "Health"
source("D:/R-project/focus-analysis/code/revindic.R")
```

## Review indicators
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis', fig.width = 10, fig.height = 7}
this.subtype <- "Health"
source("D:/R-project/focus-analysis/code/plotindic.R")
```

## Check with global recommandations

*** 
# Food Security & Nutrition
## Source Information
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis'}
this.subtype <- "Food"
source("D:/R-project/focus-analysis/code/revindic.R")
```

## Review indicators
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis', fig.width = 10, fig.height = 7}
this.subtype <- "Food"
source("D:/R-project/focus-analysis/code/plotindic.R")
```

## Check with global recommandations

*** 
# Non Food Item (NFI)
## Source Information
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis'}
this.subtype <- "NFI"
source("D:/R-project/focus-analysis/code/revindic.R")
```

## Review indicators
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis', fig.width = 10, fig.height = 7}
this.subtype <- "NFI"
source("D:/R-project/focus-analysis/code/plotindic.R")
```

## Check with global recommandations

*** 
# Shelter
## Source Information
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis'}
this.subtype <- "Shelter"
source("D:/R-project/focus-analysis/code/revindic.R")
```

## Review indicators
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis', fig.width = 10, fig.height = 7}
this.subtype <- "Shelter"
source("D:/R-project/focus-analysis/code/plotindic.R")
```

## Check with global recommandations

***
 
#  WASH
## Source Information
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis'}
this.subtype <- "WASH"
source("D:/R-project/focus-analysis/code/revindic.R")
```

## Review indicators
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis', fig.width = 10, fig.height = 7}
this.subtype <- "WASH"
source("D:/R-project/focus-analysis/code/plotindic.R")
```

## Check with global recommandations

***
#  Supply
## Source Information
 
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis'}
this.subtype <- "Supply"
source("D:/R-project/focus-analysis/code/revindic.R")
```

## Review indicators
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis', fig.width = 10, fig.height = 7}
this.subtype <- "Supply"
source("D:/R-project/focus-analysis/code/plotindic.R")
```

## Check with global recommandations

***
#  Camp Coordination and Camp Management (CCCM)
## Source Information
 
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis'}
this.subtype <- "CCCM"
source("D:/R-project/focus-analysis/code/revindic.R")
```

## Review indicators
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis', fig.width = 10, fig.height = 7}
this.subtype <- "CCCM"
source("D:/R-project/focus-analysis/code/plotindic.R")
```

## Check with global recommandations

***
 
#  Energy
## Source Information
 
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis'}
this.subtype <- "Energy"
source("D:/R-project/focus-analysis/code/revindic.R")
```

## Review indicators
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis', fig.width = 10, fig.height = 7}
this.subtype <- "Energy"
source("D:/R-project/focus-analysis/code/plotindic.R")
```

## Check with global recommandations

***
 
#  Environment
## Source Information
 
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis'}
this.subtype <- "Environment"
source("D:/R-project/focus-analysis/code/revindic.R")
```

## Review indicators
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis', fig.width = 10, fig.height = 7}
this.subtype <- "Environment"
source("D:/R-project/focus-analysis/code/plotindic.R")
```

## Check with global recommandations

***
 
#  Management
## Source Information
 
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis'}
this.subtype <- "Management"
source("D:/R-project/focus-analysis/code/revindic.R")
```

## Review indicators
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis', fig.width = 10, fig.height = 7}
this.subtype <- "Management"
source("D:/R-project/focus-analysis/code/plotindic.R")
```

## Check with global recommandations

***
 
#  External Relation
## Source Information
 
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis'}
this.subtype <- "ExtRel"
source("D:/R-project/focus-analysis/code/revindic.R")
```

## Review indicators
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis', fig.width = 10, fig.height = 7}
this.subtype <- "ExtRel"
source("D:/R-project/focus-analysis/code/plotindic.R")
```

## Check with global recommandations

***
#  Response Coordination
## Source Information
 
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis'}
this.subtype <- "Coordination"
source("D:/R-project/focus-analysis/code/revindic.R")
```

## Review indicators
```{r , echo=FALSE, warning=FALSE, message=FALSE, results='asis', fig.width = 10, fig.height = 7}
this.subtype <- "Coordination"
source("D:/R-project/focus-analysis/code/plotindic.R")
```

## Check with global recommandations

***
